Add back button on advanced settings page
