+++ lib/pages/habits/habits_page.dart
@@ -5,9 +5,10 @@ import 'package:lotti/blocs/habits/habits_cubit.dart';
 import 'package:lotti/blocs/habits/habits_state.dart';
 import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/themes/theme.dart';
+import 'package:lotti/widgets/app_bar/habit_page_app_bar.dart';
+import 'package:lotti/widgets/app_bar/sliver_title_bar.dart';
 import 'package:lotti/widgets/charts/utils.dart';
 import 'package:lotti/widgets/habits/habit_completion_card.dart';
-import 'package:lotti/widgets/habits/habit_page_app_bar.dart';
 import 'package:lotti/widgets/habits/habit_streaks.dart';
 import 'package:lotti/widgets/habits/habits_search.dart';
 import 'package:lotti/widgets/misc/timespan_segmented_control.dart';
