+++ lib/pages/settings/sliver_box_adapter_page.dart
@@ -1,16 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:lotti/themes/theme.dart';
-import 'package:lotti/widgets/habits/habit_page_app_bar.dart';
+import 'package:lotti/widgets/app_bar/sliver_title_bar.dart';
 
 class SliverBoxAdapterPage extends StatelessWidget {
   const SliverBoxAdapterPage({
     required this.child,
     required this.title,
+    this.showBackButton = false,
     super.key,
   });
 
   final Widget child;
   final String title;
+  final bool showBackButton;
 
   @override
   Widget build(BuildContext context) {
@@ -18,7 +20,11 @@ class SliverBoxAdapterPage extends StatelessWidget {
       backgroundColor: styleConfig().negspace,
       body: CustomScrollView(
         slivers: <Widget>[
-          SliverTitleBar(title),
+          SliverTitleBar(
+            title,
+            pinned: true,
+            showBackButton: showBackButton,
+          ),
           SliverToBoxAdapter(
             child: child,
           )
